pyproject.toml
[tool.poetry]
name = "pptvideo"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
click = "^7.1.2"
pdf2image = "^1.14.0"
python-pptx = "^0.6.18"
gTTS = "^2.2.2"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
pptx2video = "pptvideo.entry:p2v"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
